Add a dependent pair type
Short-term this lets us implement telescopes thanks to which we can
remove the constructor hoisting hack we've been using. It should also
make it possible to do AD of triangular for-loops, which was requested
in #587, but this is not part of this patch.

Longer-term we should be able to use it to get rid of DataConRef and
implement all ADTs as coerced sum, product and dependent pair types.